updated scripts for saving...
why do I think in complex ways... instead of comparing the strings builded, why not taking as it is? *sighn*
okay, here it is. removed the func and added a way better solution for spell components (SPELL_* constants). now they can be at every place and will be correct transmitted to later created script snippet.