update save func for components and biproducts
with this update you can save the edited components and biproducts to the script snippet shown on recipe tap.
now working on the misc tab. there will be the the level, workbench, XP and all the other things editable.
for now, only the script snipped is actualized. if you change the product in the list, it reloads it. later i will implement the saving into the array too.